Moved image processing to sofast
