+++ opencsp/app/sofast/lib/ProcessSofastFixed.py
@@ -14,7 +14,7 @@ from opencsp.common.lib.csp.MirrorPoint import MirrorPoint
 from opencsp.common.lib.deflectometry.BlobIndex import BlobIndex
 import opencsp.common.lib.deflectometry.calculation_data_classes as cdc
 from opencsp.common.lib.deflectometry.FacetData import FacetData
-import opencsp.common.lib.deflectometry.image_processing as ip
+import opencsp.app.sofast.lib.image_processing as ip
 import opencsp.app.sofast.lib.process_optics_geometry as pr
 from opencsp.common.lib.deflectometry.SlopeSolver import SlopeSolver
 from opencsp.common.lib.deflectometry.SlopeSolverData import SlopeSolverData
