+++ opencsp/app/sofast/lib/image_processing.py
