+++ opencsp/app/sofast/lib/process_optics_geometry.py
@@ -14,7 +14,7 @@ from opencsp.common.lib.deflectometry.GeometryProcessingParams import (
     GeometryProcessingParams,
 )
 from opencsp.common.lib.deflectometry.GeometryDataDebug import GeometryDataDebug
-import opencsp.common.lib.deflectometry.image_processing as ip
+import opencsp.app.sofast.lib.image_processing as ip
 from opencsp.app.sofast.lib.SpatialOrientation import SpatialOrientation
 import opencsp.app.sofast.lib.spatial_processing as sp
 from opencsp.common.lib.geometry.LoopXY import LoopXY
